Added workspace and moved see_details and select_recipient methods to workspace. Moved associated tests to workspace
