+++ specs/channel_spec.rb
@@ -25,44 +25,44 @@ describe "self.list" do
   end
 end
 
-describe "select_channel" do
-  it "selects a channel as the current recipient when user inputs channel ID" do
-    channel_url = "channels.list"
-    channel = SlackAPI::Channel.new()
+# describe "select_channel" do
+#   it "selects a channel as the current recipient when user inputs channel ID" do
+#     channel_url = "channels.list"
+#     channel = SlackAPI::Channel.new()
 
-    VCR.use_cassette("select_channel") do
-      response = channel.select_channel("CH2SKTKPC")
-      expect(response).must_equal "CH2SKTKPC"
-    end
-  end
-  it "selects a channel as the current recipient when user inputs channel name" do
-    channel = SlackAPI::Channel.new()
-    VCR.use_cassette("select_channel") do
-      response = channel.select_channel("general")
-      expect(response).must_equal "general"
-    end
-  end
-  it "raises an error if the channel is not valid" do
-    channel = SlackAPI::Channel.new()
-    VCR.use_cassette("select_channel") do
-      expect { channel.select_channel("fake channel") }.must_raise SlackAPI::SlackError
-    end
-  end
-end
+#     VCR.use_cassette("select_channel") do
+#       response = channel.select_channel("CH2SKTKPC")
+#       expect(response).must_equal "CH2SKTKPC"
+#     end
+#   end
+#   it "selects a channel as the current recipient when user inputs channel name" do
+#     channel = SlackAPI::Channel.new()
+#     VCR.use_cassette("select_channel") do
+#       response = channel.select_channel("general")
+#       expect(response).must_equal "general"
+#     end
+#   end
+#   it "raises an error if the channel is not valid" do
+#     channel = SlackAPI::Channel.new()
+#     VCR.use_cassette("select_channel") do
+#       expect { channel.select_channel("fake channel") }.must_raise SlackAPI::SlackError
+#     end
+#   end
+# end
 
-describe "see_details" do
-  it "shows details for a selected channel" do
-    channel = SlackAPI::Channel.new()
-    VCR.use_cassette("see_details") do
-      # selected_channel = channel.select_channel("CH2SKTKPC")
-      channel_details = channel.see_details("CH2SKTKPC")
+# describe "see_details" do
+#   it "shows details for a selected channel" do
+#     channel = SlackAPI::Channel.new()
+#     VCR.use_cassette("see_details") do
+#       # selected_channel = channel.select_channel("CH2SKTKPC")
+#       channel_details = channel.see_details("CH2SKTKPC")
 
-      expect(channel_details).must_be_kind_of Array
-      expect(channel_details.length).must_equal 1
-      expect(channel_details[0]).must_equal "name" => "general", "topic" => "Company-wide announcements", "member count" => 2, "id" => "CH2SKTKPC"
-    end
-  end
-end
+#       expect(channel_details).must_be_kind_of Array
+#       expect(channel_details.length).must_equal 1
+#       expect(channel_details[0]).must_equal "name" => "general", "topic" => "Company-wide announcements", "member count" => 2, "id" => "CH2SKTKPC"
+#     end
+#   end
+# end
 
 describe "send_msg" do
   it "sends a message to selected channel" do
