+++ specs/test_helper.rb
@@ -15,6 +15,7 @@ require_relative "../lib/channel.rb"
 require_relative "../lib/slack.rb"
 require_relative "../lib/user.rb"
 require_relative "../lib/recipient.rb"
+require_relative "../lib/workspace.rb"
 
 Minitest::Reporters.use! Minitest::Reporters::SpecReporter.new
